+++ Colorify/Theme/ThemeAction.cs
@@ -5,27 +5,30 @@ namespace Colorify.UI
 {
     public abstract class ThemeAction : ITheme
     {
-        protected ConsoleColor _defaultBackground { get; set; }
-        protected ConsoleColor _defaultForeground { get; set; }
+        protected bool _consoleDefault { get; set; }
+        protected ConsoleColor _consoleBackground { get; set; }
+        protected ConsoleColor _consoleForeground { get; set; }
 
         public Dictionary<string, Color> _colors { get; set; }
 
         public abstract void SetColors();
 
-        public ThemeAction()
+        public abstract void SetComponents();
+
+        public void ResetColor()
         {
-            _defaultBackground = Console.BackgroundColor;
-            _defaultForeground = Console.ForegroundColor;
-            SetColors();
+            Console.ResetColor();
+            _consoleBackground = Console.BackgroundColor;
+            _consoleForeground = Console.ForegroundColor;
         }
+
         public Color AddColor(ConsoleColor? background, ConsoleColor? foreground)
         {
             var color = new Color(
-                background ?? _defaultBackground,
-                foreground ?? _defaultForeground
+                background ?? _consoleBackground,
+                foreground ?? _consoleForeground
             );
             return color;
         }
     }
 }
-