+++ Colorify/Theme/ThemeLight.cs
@@ -9,14 +9,34 @@ namespace Colorify.UI
         {
             get
             {
-                return new ThemeLight();
+                return new ThemeLight(false);
             }
         }
     }
 
-    public sealed class ThemeLight : ThemeAction
+    public class ThemeLight : ThemeAction
     {
+        public ThemeLight(bool consoleDefault = true)
+        {
+            base._consoleDefault = consoleDefault;
+            SetColors();
+            SetComponents();
+        }
+
         public override void SetColors()
+        {
+            if (base._consoleDefault)
+            {
+                ResetColor();
+            }
+            else
+            {
+                base._consoleBackground = ConsoleColor.White;
+                base._consoleForeground = ConsoleColor.Black;
+            }
+        }
+
+        public override void SetComponents()
         {
             var colors = new Dictionary<string, Color>();
             colors.Add("text-default", AddColor(null, null));
@@ -33,7 +53,7 @@ namespace Colorify.UI
             colors.Add("text-info", AddColor(null, ConsoleColor.DarkCyan));
             colors.Add("bg-success", AddColor(ConsoleColor.DarkGreen, ConsoleColor.White));
             colors.Add("bg-info", AddColor(ConsoleColor.DarkCyan, ConsoleColor.White));
-            _colors = colors;
+            base._colors = colors;
         }
     }
 }